xmodmap changes for ergodox changes
The ErgoDox is now programmed to send a "Super" keycode instead of "Alt"
because Qt has an unconfigurable behavior of stealing input focus when Alt is
pressed (it assumes you want to access the menu), which then proceeds to steal
shortcuts from terminal programs like tmux.

QMK has no option to send an actual Meta--and I'm not sure if Meta is
even defined by the HID standard--so I've configured the keycap to send
Super because it can then be reassigned to Meta in X where it can
be added to mod1 and interpreted as Alt! This means xev(1) shows the
keysym as "Meta," but applications continue to interpret it as "Alt"
(via mod1)... yet Qt is smart enough to recognize that it's *not really*
Alt, and alas it is no longer stealing my fucking input focus or bindings.

(The input focus behavior is really a bad one, IMO: After you've pressed
 Alt and focus is now on the menu bar instead of the inner window, it's
 not visually obvious and your subsequent keystrokes to your editor, shell,
 Slack, or whatever are effectively lost as you wonder where the hell
 focus went.)

While KWin has continued to interpret the new key as Alt (KWin refers to
Super as "Meta," FWIW, as does the rest of the modern Internet apparently),
it triggered one of those hacky bugs internally that broke Alt-Tab; after
nearly endless trial and error trying to resolve the situation, I have
relented and rebound the app switcher to Ctrl-Esc instead (which is more
comfortable on my layout anyhow).

Hmph.  Qt.